Added splash screen on start and game start game over logic.
Modified game_reset
Added start_game
Modified on_draw
Modified on_update

+++ asteroids.py
@@ -4,6 +4,8 @@ import utils
 from pyglet.window import key
 from utils import group_collide, group_group_collide
 from sprites import MovingSprite, PlayerSprite
+from pyglet.sprite import Sprite
+
 # Global constants
 WIDTH = 800
 HEIGHT = 600
@@ -33,10 +35,13 @@ class Window(pyglet.window.Window):
         ship_sequence = pyglet.image.load('static/images/double_ship.png')
         ship_imgs = pyglet.image.ImageGrid(ship_sequence, 1, 2)
         utils.center_image_grid_anchors(ship_imgs)
+        self.splash_img = pyglet.image.load('static/images/splash.png') 
 
         # Sounds
         thruster_snd = pyglet.media.load('static/sounds/rocket.ogg', streaming=False)
         self.explosion_snd = pyglet.media.load('static/sounds/explosion.ogg', streaming=False)
+        self.background_music = pyglet.media.load('static/sounds/space1.mp3', streaming=False)
+        self.background_music.play()
 
         # Sprite Groups
         self.rock_group = set()
@@ -44,6 +49,7 @@ class Window(pyglet.window.Window):
 
         # Spites 
         self.ship = PlayerSprite(ship_imgs, thruster_snd, 400, 250, 0, 0, 270, 35, self.ships, self.missiles)
+        self.splash = Sprite(self.splash_img, 200, 125)
 
         # Screen Text
         self.text_lives = pyglet.text.Label('Lives=' + str(self.lives),
@@ -58,13 +64,14 @@ class Window(pyglet.window.Window):
         self.key_ups = {key.UP:self.decel, key.LEFT:self.right, key.RIGHT:self.left}
 
         pyglet.clock.schedule_interval(self.update, 1/60.0)  # update at 60Hz
-    def game_over():
+    def game_reset(self):
         self.started = False
         rock_group = set()
         # sountrack.pause()
         self.lives = 3
         self.score = 0
         # intro screen 
+        # self.splash.draw()
 
     def display_score(self):
         self.text_lives = pyglet.text.Label('Lives=' + str(self.lives),
@@ -73,7 +80,10 @@ class Window(pyglet.window.Window):
         self.text_score = pyglet.text.Label('Score=' + str(self.score),
                                             font_name='Times New Roman',
                                             font_size=36, x=10, y=60)
-
+    def start_game(self):
+        #Function to start game on initial mouse click
+        self.splash.delete()
+        self.started = True
 
     def accel(self):
         self.ship.thrusters = True
@@ -89,6 +99,9 @@ class Window(pyglet.window.Window):
     
     def fire(self):
         self.ship.shoot(self.missile_group)
+
+    def on_mouse_press(self, x, y, button, modifiers):
+        self.start_game()
 
     def on_key_press(self, symbol, modifiers):
         for key in self.key_downs:
@@ -123,29 +136,36 @@ class Window(pyglet.window.Window):
         self.missiles.draw()
         self.text_lives.draw()
         self.text_score.draw()
+        if not self.started:
+           self.splash = Sprite(self.splash_img, 200, 125)
+           self.splash.draw()
 
     def update(self, dt):
-        self.trigger_put_rock()
-        for rock in self.rock_group:
-            rock.update(WIDTH, HEIGHT) 
-        self.ship.update(WIDTH, HEIGHT) 
+        if self.started:
+            self.trigger_put_rock()
+            for rock in self.rock_group:
+                rock.update(WIDTH, HEIGHT) 
+            self.ship.update(WIDTH, HEIGHT) 
 
-        local_missiles = set(self.missile_group)
-        for missile in local_missiles:
-            if not missile.update(WIDTH, HEIGHT):
-                self.missile_group.remove(missile)
-                missile.delete()
+            local_missiles = set(self.missile_group)
+            for missile in local_missiles:
+                if not missile.update(WIDTH, HEIGHT):
+                    self.missile_group.remove(missile)
+                    missile.delete()
 
-        if group_collide(self.rock_group, self.ship):
+            if group_collide(self.rock_group, self.ship):
                 self.lives -= 1
                 self.explosion_snd.play()
 
-        if group_group_collide(self.missile_group, self.rock_group):
+            if group_group_collide(self.missile_group, self.rock_group):
                 self.score += 10
                 self.difficulty  += self.score + 50
                 self.explosion_snd.play()
 
-        self.display_score()
+            self.display_score()
+        
+        if self.lives < 0:
+            self.game_reset()
 
 
 def main():
